feat: 调整react前端样式
