+++ laokou-service/laokou-logstash/laokou-logstash-adapter/src/main/java/org/laokou/logstash/consumer/TraceLogConsumer.java
@@ -19,21 +19,13 @@ package org.laokou.logstash.consumer;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.laokou.common.core.utils.IdGenerator;
-import org.laokou.common.core.utils.MapUtil;
-import org.laokou.common.i18n.utils.JacksonUtil;
-import org.laokou.common.i18n.utils.StringUtil;
 import org.laokou.logstash.common.support.TraceLogStorage;
-import org.laokou.logstash.gateway.database.dataobject.TraceLogIndex;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.support.Acknowledgment;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
-import java.util.stream.Collectors;
 
 /**
  * @author laokou
@@ -48,45 +40,15 @@ public class TraceLogConsumer {
 	@KafkaListener(topics = "laokou_trace_topic", groupId = "laokou_trace_consumer_group")
 	public CompletableFuture<Void> kafkaConsumer(List<String> messages, Acknowledgment ack) {
 		try {
-			Map<String, Object> dataMap = messages.parallelStream()
-				.map(this::getTraceLogIndex)
-				.filter(Objects::nonNull)
-				.collect(Collectors.toMap(TraceLogIndex::getId, traceLogIndex -> traceLogIndex));
-			if (MapUtil.isNotEmpty(dataMap)) {
-				return traceLogStorage.batchSave(dataMap);
-			}
+			return traceLogStorage.batchSave(messages);
 		}
 		catch (Exception e) {
 			log.error("分布式链路写入失败，错误信息：{}", e.getMessage());
+			return null;
 		}
 		finally {
 			ack.acknowledge();
 		}
-		return null;
-	}
-
-	private TraceLogIndex getTraceLogIndex(String str) {
-		try {
-			TraceLogIndex traceLogIndex = JacksonUtil.toBean(str, TraceLogIndex.class);
-			String traceId = traceLogIndex.getTraceId();
-			String spanId = traceLogIndex.getSpanId();
-			if (isTraceLog(traceId, spanId)) {
-				traceLogIndex.setId(String.valueOf(IdGenerator.defaultSnowflakeId()));
-				return traceLogIndex;
-			}
-		}
-		catch (Exception ex) {
-			log.error("分布式链路日志JSON转换失败，错误信息：{}", ex.getMessage());
-		}
-		return null;
-	}
-
-	private boolean isTraceLog(String traceId, String spanId) {
-		return isTraceLog(traceId) && isTraceLog(spanId);
-	}
-
-	private boolean isTraceLog(String str) {
-		return StringUtil.isNotEmpty(str) && !str.startsWith("${") && !str.endsWith("}");
 	}
 
 }
