+++ laokou-service/laokou-logstash/laokou-logstash-infrastructure/src/main/java/org/laokou/logstash/common/support/TraceLogElasticsearchStorage.java
@@ -18,13 +18,17 @@
 package org.laokou.logstash.common.support;
 
 import lombok.RequiredArgsConstructor;
+import org.laokou.common.core.utils.MapUtil;
 import org.laokou.common.core.utils.ThreadUtil;
 import org.laokou.common.elasticsearch.template.ElasticsearchTemplate;
 import org.laokou.logstash.gateway.database.dataobject.TraceLogIndex;
 
+import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
+import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 public class TraceLogElasticsearchStorage extends AbstractTraceLogStorage {
@@ -32,16 +36,25 @@ public class TraceLogElasticsearchStorage extends AbstractTraceLogStorage {
 	private final ElasticsearchTemplate elasticsearchTemplate;
 
 	@Override
-	public CompletableFuture<Void> batchSave(Map<String, Object> map) {
-		try (ExecutorService executor = ThreadUtil.newVirtualTaskExecutor()) {
-			return elasticsearchTemplate.asyncCreateIndex(getIndexName(), TRACE_INDEX, TraceLogIndex.class, executor)
-				.thenAcceptAsync(res -> elasticsearchTemplate.asyncBulkCreateDocument(getIndexName(), map, executor),
-						executor);
+	public CompletableFuture<Void> batchSave(List<String> messages) {
+		Map<String, Object> dataMap = messages.parallelStream()
+			.map(this::getTraceLogIndex)
+			.filter(Objects::nonNull)
+			.collect(Collectors.toMap(TraceLogIndex::getId, traceLogIndex -> traceLogIndex));
+		if (MapUtil.isNotEmpty(dataMap)) {
+			try (ExecutorService executor = ThreadUtil.newVirtualTaskExecutor()) {
+				return elasticsearchTemplate
+					.asyncCreateIndex(getIndexName(), TRACE_INDEX, TraceLogIndex.class, executor)
+					.thenAcceptAsync(
+							res -> elasticsearchTemplate.asyncBulkCreateDocument(getIndexName(), dataMap, executor),
+							executor);
+			}
 		}
+		return null;
 	}
 
 	@Override
-	public CompletableFuture<Void> save(Object obj) {
+	public CompletableFuture<Void> save(String obj) {
 		throw new UnsupportedOperationException();
 	}
