+++ laokou-service/laokou-logstash/laokou-logstash-infrastructure/src/main/java/org/laokou/logstash/common/support/TraceLogLokiStorage.java
@@ -17,19 +17,19 @@
 
 package org.laokou.logstash.common.support;
 
-import java.util.Map;
+import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
 public class TraceLogLokiStorage extends AbstractTraceLogStorage {
 
 	@Override
-	public CompletableFuture<Void> batchSave(Map<String, Object> map) {
+	public CompletableFuture<Void> batchSave(List<String> messages) {
 		return null;
 	}
 
 	@Override
-	public CompletableFuture<Void> save(Object obj) {
-		throw new UnsupportedOperationException();
+	public CompletableFuture<Void> save(String obj) {
+		return null;
 	}
 
 }
